tests/requirements.txt

###### Requirements without Version Specifiers ######`
wheel
base45
jsonschema
jsonref
filecache
pyzbar
Pillow
python-dateutil
pytest-md-report
pytest-html
requests

###### Requirements with Version Specifiers ######`
cose == 0.9.dev7